platform/apps/gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
asyncpg==0.29.0
clickhouse-driver==0.2.6
prometheus-client==0.19.0
python-multipart==0.0.6
httpx==0.26.0

